Add network status command and refactor system event handling
- Introduced `HandleNetworkStatusCommand` to print network registration and activity status.
- Updated `uart_cmd_table` to include the new `netstatus` command.
- Refactored event handling in `system.c` to improve clarity and maintainability.
- Enhanced global variable management for cell information and system status.
