+++ app/src/gps_tracker.c
@@ -1,50 +1,20 @@
-#include <string.h>
 #include <stdio.h>
 #include <api_os.h>
 #include <api_gps.h>
-#include <api_info.h>
 #include <api_event.h>
 #include <api_network.h>
-#include <api_fs.h>
-#include <api_lbs.h>
-#include <api_hal_uart.h>
 #include <api_hal_pm.h>
 
-#include "math.h"
-#include "time.h"
-#include "assert.h"
-#include "buffer.h"
+#include "system.h"
 #include "gps.h"
 #include "gps_parse.h"
-#include "sms_service.h"
-
-#include "system.h"
 #include "gps_tracker.h"
+#include "sms_service.h"
 #include "config_store.h"
 #include "config_commands.h"
-#include "led_handler.h"
 #include "network.h"
 #include "utils.h"
 #include "debug.h"
-
-#define MAIN_TASK_STACK_SIZE      (2048 * 2)
-#define MAIN_TASK_PRIORITY        (0)
-#define MAIN_TASK_NAME            "GPS Tracker"
-
-#define REPORTING_TASK_STACK_SIZE (2048 * 2)
-#define REPORTING_TASK_PRIORITY   (0)
-#define REPORTING_TASK_NAME       "Reporting Task"
-
-HANDLE gpsMainTaskHandle   = NULL;
-HANDLE reportingTaskHandle = NULL;
-
-uint8_t systemStatus = 0;
-uint8_t g_RSSI = 0;
-char    g_cellInfo[128] = "\0";
-// Provide last known position for SMS service
-float   g_last_latitude = 0.0f;
-float   g_last_longitude = 0.0f;
-
 
 SSL_Config_t SSLconfig = {
     .caCert          = NULL,
@@ -76,7 +46,7 @@ SSL_Config_t SSLconfig = {
 Network_PDP_Context_t NetContextArr[2];      // Two-element array: [0]=real APN, [1]=dummy APN
 static bool apn_workaround_pending = false;  
 
-static void apnWorkaround_init(void) 
+ void apnWorkaround_init(void) 
 {
     // set Workaround state for APN re-activation
     apn_workaround_pending = false;
@@ -94,7 +64,7 @@ static void SetApnContext() {
     strncpy(NetContextArr[0].userPasswd, g_ConfigStore.apn_pass, sizeof(NetContextArr[0].userPasswd)-1);
 }
 
-bool AttachActivate()
+bool gsm_AttachActivate()
 {
     uint8_t status;
     bool ret = Network_GetAttachStatus(&status);
@@ -145,179 +115,10 @@ bool AttachActivate()
     return true;
 }
 
-uint8_t csq_to_percent(int csq)
-{
-    if (csq < 0 || csq == 99) {
-        // 99 means unknown or undetectable signal
-        return 0;
-    }
-    if (csq > 31) {
-        csq = 31;  // Clamp max value
-    }
-    // Scale 0–31 to 0–100%
-    return (csq * 100) / 31;
-}
-
-void processNetworkCellInfo(Network_Location_t* loc, int number)
-{
-    g_cellInfo[0] = '\0';
-    if (number<=0) return;
-    snprintf(g_cellInfo,  sizeof(g_cellInfo), "%u%u%u,%u%u%u,%u,%u,%d",
-             loc->sMcc[0], loc->sMcc[1], loc->sMcc[2], loc->sMnc[0], loc->sMnc[1], loc->sMnc[2], loc->sLac, loc->sCellID, loc->iRxLev);
-}
-
-static void format_size(char* buf, size_t bufsize, int value, const char* label) {
-    if (value >= 1024*1024*1024) {
-        float gb = value / (1024.0f * 1024.0f * 1024.0f);
-        snprintf(buf, bufsize, "%s%.3f GB", label, gb);
-    } else if (value >= 1024*1024) {
-        float mb = value / (1024.0f * 1024.0f);
-        snprintf(buf, bufsize, "%s%.1f MB", label, mb);
-    } else if (value >= 1024) {
-        float kb = value / 1024.0f;
-        snprintf(buf, bufsize, "%s%.1f KB", label, kb);
-    } else {
-        snprintf(buf, bufsize, "%s%d Bytes", label, value);
-    }
-}
-
-void FsInfoTest()
-{
-    API_FS_INFO fsInfo;
-    int sizeUsed = 0, sizeTotal = 0;
-    char used_buf[32], total_buf[32];
-
-    if(API_FS_GetFSInfo(FS_DEVICE_NAME_FLASH, &fsInfo) < 0)
-    {
-        LOGE("Get Int Flash device info fail!");
-    }
-    sizeUsed  = fsInfo.usedSize;
-    sizeTotal = fsInfo.totalSize;
-    format_size(used_buf, sizeof(used_buf), sizeUsed, "");
-    format_size(total_buf, sizeof(total_buf), sizeTotal, "");
-    LOGI("Int Flash used: %s, total size: %s", used_buf, total_buf);
-
-    if(API_FS_GetFSInfo(FS_DEVICE_NAME_T_FLASH, &fsInfo) < 0)
-    {
-        LOGE("Get Ext Flash device info fail!");
-        return;
-    }
-    sizeUsed  = fsInfo.usedSize;
-    sizeTotal = fsInfo.totalSize;
-    format_size(used_buf, sizeof(used_buf), sizeUsed, "");
-    format_size(total_buf, sizeof(total_buf), sizeTotal, "");
-    LOGI("Ext Flash used: %s, total size: %s", used_buf, total_buf);
-}
-
-void EventHandler(API_Event_t* pEvent)
-{
-    switch(pEvent->id)
-    {
-        case API_EVENT_ID_NO_SIMCARD:
-            GSM_ACTIVE_OFF();
-            LOGE("no sim card %d !", pEvent->param1);
-            break;
-        case API_EVENT_ID_SIMCARD_DROP:
-            GSM_ACTIVE_OFF();
-            LOGE("sim card %d drop !",pEvent->param1);
-            break;
-        case API_EVENT_ID_NETWORK_REGISTER_DENIED:
-            GSM_REGISTERED_OFF();
-            LOGE("network register denied");
-            break;
-        case API_EVENT_ID_NETWORK_REGISTER_NO:
-            GSM_REGISTERED_OFF();
-            LOGE("network register no");
-            break;
-
-        case API_EVENT_ID_NETWORK_REGISTERED_HOME:
-        case API_EVENT_ID_NETWORK_REGISTERED_ROAMING:
-            GSM_REGISTERED_ON();    
-            LOGW("network register success");
-            AttachActivate();
-            break;
-
-        case API_EVENT_ID_NETWORK_ATTACHED:
-            GSM_ACTIVE_OFF(); 
-            LOGW("network attach success");
-            break;
-
-        case API_EVENT_ID_NETWORK_ACTIVATED:
-            GSM_ACTIVE_ON();
-            LOGW("network activate success");
-            break;
-
-        case API_EVENT_ID_NETWORK_ACTIVATE_FAILED:
-            LOGE("network activate failed");
-            break;
-
-        case API_EVENT_ID_NETWORK_DEACTIVED:
-            GSM_ACTIVE_OFF(); 
-            LOGE("network deactived");
-            break;
-
-        case API_EVENT_ID_NETWORK_ATTACH_FAILED:
-            GSM_ACTIVE_OFF(); 
-            LOGE("network attach failed");
-            break;
-
-        case API_EVENT_ID_NETWORK_DETACHED:
-            LOGE("network detached");
-            break;
-
-        case API_EVENT_ID_SIGNAL_QUALITY:
-            g_RSSI = csq_to_percent(pEvent->param1);
-            break;
-
-        case API_EVENT_ID_NETWORK_CELL_INFO:
-            processNetworkCellInfo((Network_Location_t*)pEvent->pParam1, pEvent->param1);
-            break;
-
-        case API_EVENT_ID_SYSTEM_READY:
-            LOGW("system initialize complete");
-            INITIALIZED_ON();
-            break;
-
-        case API_EVENT_ID_SMS_RECEIVED:
-            HandleSmsReceived(pEvent);
-            break;
-
-        case API_EVENT_ID_GPS_UART_RECEIVED:
-            if (g_ConfigStore.gps_logging)
-                LOGD("received GPS data, length:%d, data:\r\n%s",pEvent->param1,pEvent->pParam1);
-
-            GPS_STATUS_ON();
-            GPS_Update(pEvent->pParam1, pEvent->param1);
-
-            GPS_Info_t* gpsInfo = Gps_GetInfo();
-            if (gpsInfo->rmc.valid) {
-                GPS_FIX_ON();
-                g_last_latitude = minmea_tocoord(&gpsInfo->rmc.latitude);
-                g_last_longitude = minmea_tocoord(&gpsInfo->rmc.longitude);
-            } else {
-                GPS_FIX_OFF();
-            }
-
-            break;
-        
-        case API_EVENT_ID_UART_RECEIVED:
-            if(pEvent->param1 == UART1)
-            {
-                uint8_t data[(pEvent->param2 + 1)];
-                data[pEvent->param2] = 0;
-                memcpy(data,pEvent->pParam1,pEvent->param2);
-                    HandleUartCommand(data);
-            }
-            break;
-        default:
-            break;
-    }
-}
-
 uint8_t requestBuffer[400];
 uint8_t responseBuffer[1024];
 
-void gps_trackingTask(void *pData)
+void gps_trackerTask(void *pData)
 {
     while (!IS_INITIALIZED() || !IS_GSM_ACTIVE()) OS_Sleep(2000);
 
@@ -364,24 +165,14 @@ void gps_trackingTask(void *pData)
     {
         GPS_Info_t* gpsInfo = Gps_GetInfo();
         uint32_t    loop_start = time(NULL);
-        if(IS_GPS_STATUS_ON() && (gpsInfo->rmc.valid))
+        if(IS_GPS_STATUS_ON() && (IS_GPS_FIX()))
         {
-            if(!Network_GetCellInfoRequst()) {
-                g_cellInfo[0] = '\0';
-                LOGE("network get cell info fail");
-            }
-
             time_t gps_timestamp = mk_time(&gpsInfo->rmc.date, &gpsInfo->rmc.time);
-
-            // convert coordinates ddmm.mmmm to a floating point DD.DDD value in degree(°) 
-            float latitude  = minmea_tocoord(&gpsInfo->rmc.latitude);
-            float longitude = minmea_tocoord(&gpsInfo->rmc.longitude);
             
             // convert other data a floating point 
             float speed     = minmea_tofloat(&gpsInfo->rmc.speed);
             float bearing   = minmea_tofloat(&gpsInfo->rmc.course);
             float altitude  = minmea_tofloat(&gpsInfo->gga.altitude);
-            
             float accuracy0 = minmea_tofloat(&gpsInfo->gsa[0].hdop);
 
             float gps_uere = g_ConfigStore.gps_uere;
@@ -413,7 +204,7 @@ void gps_trackingTask(void *pData)
 
             snprintf(requestBuffer, sizeof(requestBuffer),
                      "id=%s&valid=%d&timestamp=%d&lat=%f&lon=%f&speed=%1.f&bearing=%.1f&altitude=%.1f&accuracy=%.1f%s&batt=%d",
-                     device_name, gpsInfo->rmc.valid, gps_timestamp, latitude, longitude, speed, bearing, altitude, accuracy, responseBuffer, percent);
+                     device_name, gpsInfo->rmc.valid, gps_timestamp, g_last_latitude, g_last_longitude, speed, bearing, altitude, accuracy, responseBuffer, percent);
             requestBuffer[sizeof(requestBuffer) - 1] = '\0';
 
             if(IS_GSM_ACTIVE())
@@ -449,64 +240,6 @@ void gps_trackingTask(void *pData)
         uint32_t desired_interval = 10; // target loop period in seconds (e.g., 10 seconds)
 
         if (loop_duration > desired_interval) loop_duration = desired_interval;
-        OS_Sleep((desired_interval - loop_duration)*1000);
+        OS_Sleep((desired_interval - loop_duration) * 1000);
     }
-}
-
-void gps_MainTask(void *pData)
-{
-    // UART1 to print user logs
-    UART_Config_t config = {
-        .baudRate = UART_BAUD_RATE_115200,
-        .dataBits = UART_DATA_BITS_8,
-        .stopBits = UART_STOP_BITS_1,
-        .parity   = UART_PARITY_NONE,
-        .rxCallback = NULL,
-        .useEvent   = true
-    };
-
-    TIME_SetIsAutoUpdateRtcTime(true);
-
-    PM_PowerEnable(POWER_TYPE_VPAD,true);
-    UART_Init(UART1, config);
-    
-    UART_Printf("Initialization ...\r\n");
-    apnWorkaround_init();
-    ConfigStore_Init();
-    FsInfoTest();    
-    LED_init();
-    LED_cycle_start(gpsMainTaskHandle);
-    GPS_Init();
-    SMSInit();
-
-    reportingTaskHandle = OS_CreateTask(
-        gps_trackingTask, NULL, NULL,
-        REPORTING_TASK_STACK_SIZE,
-        REPORTING_TASK_PRIORITY, 
-        0, 0, REPORTING_TASK_NAME);
-    
-    //Dispatch loop
-    while(1)
-    {
-        API_Event_t* event = NULL;
-
-        if(OS_WaitEvent(gpsMainTaskHandle, (void**)&event, OS_TIME_OUT_WAIT_FOREVER))
-        {
-            EventHandler(event);
-            OS_Free(event->pParam1);
-            OS_Free(event->pParam2);
-            OS_Free(event);
-        }
-    }
-}
-
-void app_Main(void)
-{
-    gpsMainTaskHandle = OS_CreateTask(
-        gps_MainTask, NULL, NULL,
-        MAIN_TASK_STACK_SIZE, 
-        MAIN_TASK_PRIORITY,
-        0, 0, MAIN_TASK_NAME);
-
-    OS_SetUserMainHandle(&gpsMainTaskHandle);
 }
