+++ app/src/network.h
@@ -25,4 +25,25 @@ int Http_Post(SSL_Config_t *sslConfig,
               char*         retBuffer,
               int           retBufferSize);
 
+/**
+ * @brief Processes the network cell information and updates the global cell info string.
+ * 
+ * This function formats the provided network location data into a string representation
+ * and stores it in the global variable `g_cellInfo`.
+ * 
+ * @param loc Pointer to a Network_Location_t structure containing cell information.
+ * @param number The number of cells to process (used for validation).
+ */
+void processNetworkCellInfo(Network_Location_t* loc, int number);
+
+/*
+ * @brief Global variable to store cell information.
+ *
+ * This variable is used to hold the formatted cell information string
+ * that can be accessed by other parts of the application.  
+ */
+extern char g_cellInfo[];
+
+void network_info_start(HANDLE taskHandle);
+
 #endif
